Write code to pass first reducer test
